refactor: feign client 채널 로직 분리
