+++ src/backend/service-server/src/main/java/com/bbebig/serviceserver/global/feign/FeignController.java
@@ -2,6 +2,8 @@ package com.bbebig.serviceserver.global.feign;
 
 import com.bbebig.commonmodule.clientDto.ServiceFeignResponseDto.*;
 import com.bbebig.commonmodule.global.response.code.CommonResponse;
+import com.bbebig.commonmodule.redis.domain.ChannelLastInfo;
+import com.bbebig.serviceserver.channel.service.ChannelService;
 import com.bbebig.serviceserver.server.service.ServerService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -24,6 +26,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class FeignController {
 
 	private final ServerService serverService;
+	private final ChannelService channelService;
 
 	@Operation(summary = "서버에 속해있는 채널 목록 조회 (For FeignClient)", description = "서버에 속해있는 채널 목록을 조회합니다.")
 	@ApiResponses(value = {
@@ -64,7 +67,7 @@ public class FeignController {
 		return serverService.getMemberServerList(memberId);
 	}
 
-	@Operation(summary = "서버별 채널 마지막 방문 정보 조회 (For FeignClient", description = "서버별 채널 마지막 방문 정보를 조회합니다.")
+	@Operation(summary = "서버별 채널 마지막 방문 정보 조회 (For FeignClient)", description = "서버별 채널 마지막 방문 정보를 조회합니다.")
 	@ApiResponses(value = {
 			@ApiResponse(responseCode = "200", description = "채널 마지막 방문 정보 조회 성공", useReturnTypeSchema = true),
 			@ApiResponse(responseCode = "400", description = "", content = @Content)
@@ -76,5 +79,16 @@ public class FeignController {
 	) {
 		log.info("[Service] 서버별 채널 마지막 방문 정보 조회 요청: serverId = {}, memberId = {}", serverId, memberId);
 		return serverService.getServerChannelLastInfoForApi(memberId, serverId);
+	}
+
+	@Operation(summary = "채널 마지막 방문 정보 조회(For FeignClient)", description = "채널 마지막 방문 정보를 조회합니다.")
+	@ApiResponses(value = {
+			@ApiResponse(responseCode = "200", description = "채널 마지막 방문 정보 조회 성공", useReturnTypeSchema = true),
+			@ApiResponse(responseCode = "400", description = "", content = @Content)
+	})
+	@GetMapping("/channels/{channelId}/lastInfo/member/{memberId}")
+	public CommonResponse<ChannelLastInfo> getChannelLastInfo(@PathVariable Long channelId, @PathVariable Long memberId) {
+		log.info("[Service] 채널 마지막 방문 정보 조회 요청: channelId = {}, memberId = {}", channelId, memberId);
+		return CommonResponse.onSuccess(channelService.getChannelLastInfo(channelId, memberId));
 	}
 }
